+++ src/Bot.ts
@@ -1,4 +1,10 @@
-import { Client, Collection, IntentsBitField } from 'discord.js'
+import {
+  Client,
+  Collection,
+  Events,
+  IntentsBitField,
+  Interaction,
+} from 'discord.js'
 
 import commands from './commands/index'
 import { Environment } from './config'
@@ -43,6 +49,10 @@ export class Bot {
   }
 
   loadHandlers() {
+    this.client.once(Events.ClientReady, () => this.onReady())
+    this.client.once(Events.InteractionCreate, (interaction) =>
+      this.onInteractionCreate(interaction),
+    )
     this.loadPlugins([...eventPlugins, ...featurePlugins])
     this.loadCommandHandlers(commands)
   }
@@ -75,6 +85,65 @@ export class Bot {
   private loadCommandHandlers(handlers: CommandHandlerConfig[]) {
     for (const handler of handlers) {
       this.commands.set(handler.data.name, handler)
+    }
+  }
+
+  private async onReady() {
+    const { client, commands } = this
+    console.log(`[READY] Now online as ${client.user?.tag}.`)
+    const commands_data = [...commands.values()].map((command) => command.data)
+
+    // Set guild commands
+    try {
+      const guild = client.guilds.cache.get(Environment.GUILD_ID)
+      if (!guild) {
+        throw new Error(`Guild ${Environment.GUILD_ID} not found`)
+      }
+      await guild.commands.set(commands_data)
+      console.info(
+        `[READY] ${commands.size} guild commands registered on ${guild.name}`,
+      )
+    } catch (error) {
+      console.error('[READY] Unable to set guild commands:', error)
+    }
+
+    // Clear global commands
+    try {
+      const commands = await client.application?.commands.fetch()
+      for (const command of commands?.values() || []) {
+        await command.delete()
+        console.info(`[READY] Deleted global command ${command.name}`)
+      }
+    } catch (error) {
+      console.error('[READY] Unable to clear application commands:', error)
+    }
+  }
+
+  private async onInteractionCreate(interaction: Interaction) {
+    const { commands } = this
+    if (interaction.isCommand()) {
+      const botContext = this.createBotContext()
+      const commandName = interaction.commandName
+      const command = commands.get(commandName)
+      if (!command) return
+      try {
+        if (!command.disableAutoDeferReply) {
+          await interaction.deferReply({ ephemeral: command.ephemeral })
+        }
+      } catch (error) {
+        console.error(`[COMMAND: ${commandName}] Unable to defer reply:`, error)
+        return
+      }
+      try {
+        console.info(
+          `[COMMAND: ${commandName}] Invoked by ${interaction.user.tag} (${interaction.user.id})`,
+        )
+        await command.execute(botContext, interaction)
+      } catch (error) {
+        console.error(`[COMMAND: ${commandName}] Unable to execute:`, error)
+        if (command.disableAutoDeferReply) return
+        await interaction.deleteReply()
+      }
     }
   }
 }
